Now returns the dieType and correct number of rolls based on request
A response to a roll request now returns the die type and a list of
'results'. The list contains the number of results that matches with the
number of rolls requested.

The results an simply an incrementing number starting at one and thus
can exceed the die type (eg. a 6 sided die will currently return roll
results higher than 6 when more than 6 rolls are requested).