DiceRollService now calls DiceRollStats
When a request to the service is made to roll for a die, the service
after generating the rolls, calls out to the DiceRollStats with the
results of the rolls to gather the stats regarding the roll(s). It then
adds these to the DieRoll object that was passed in as part of the
request to generate die rolls.