Refactor the DiceRollController to return the DieRoll as JSON.
Instead of crafting a string response in the shape of JSON, we let
Spring boot take care of converting the DieRoll POJO into JSON.

This required refactors to the test for said controller. It also results
in the stats calculated by DieRollStats being included in the response
JSON.

Backfilling tests for this will be required.