Add SecureRandom to DiceRollService to introduce random rolls
A new instance of SecureRandom is created for each roll.
Due to how DiceRollStats operates, the results of the roll are placed
into ascending order rather than the order they were rolled. Ascending
order is needed in order to determine the median value, but the roll
results returned should reflect the order in which result were obtained.
Need to correct this behaviour in future.